requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
gradio
pytest
